feat(prechecks): add prechecks suite package
